Revert "Merge pull request #237 from swirlai/1.11"
This reverts commit 09f16e08c042f601fa8cf31c417075fc9f044364, reversing
changes made to a8c83277d5564ee2d6c4d3d3ba96064133527341.
