--- SearchProviders/m365_outlook_calendar.json
@@ -1,26 +1,0 @@
-{
-    "name": "M365 Outlook Calendar",
-    "shared": true,
-    "active": true,
-    "default": true,
-    "connector": "M365OutlookCalendar",
-    "url": "",
-    "query_template": "{url}",
-    "query_processor": "",
-    "query_processors": [
-        "AdaptiveQueryProcessor"
-    ],
-    "query_mappings": "NOT=true,NOT_CHAR=-",
-    "result_processor": "",
-    "result_processors": [
-        "MappingResultProcessor"
-    ],
-    "response_mappings": "",
-    "result_mappings": "title=resource.subject,body=summary,date_published=resource.start.dateTime,NO_PAYLOAD",
-    "results_per_query": 10,
-    "credentials": "",
-    "eval_credentials": "",
-    "tags": [
-        "mgraph"
-    ]
-}
