--- SearchProviders/m365_sharepoint.json
@@ -1,26 +1,0 @@
-{
-    "name": "M365 SharePoint",
-    "shared": true,
-    "active": true,
-    "default": true,
-    "connector": "M365SharepointSites",
-    "url": "",
-    "query_template": "{url}",
-    "query_processor": "",
-    "query_processors": [
-        "AdaptiveQueryProcessor"
-    ],
-    "query_mappings": "NOT=true,NOT_CHAR=-",
-    "result_processor": "",
-    "result_processors": [
-        "MappingResultProcessor"
-    ],
-    "response_mappings": "",
-    "result_mappings": "title=resource.name,body=resource.description,date_published=resource.createdDateTime,NO_PAYLOAD",
-    "results_per_query": 10,
-    "credentials": "",
-    "eval_credentials": "",
-    "tags": [
-        "mgraph"
-    ]
-}
