--- SearchProviders/microsoft_teams.json
@@ -1,26 +1,0 @@
-{
-    "name": "Microsoft Teams",
-    "shared": true,
-    "active": true,
-    "default": true,
-    "connector": "MicrosoftTeams",
-    "url": "",
-    "query_template": "{url}",
-    "query_processor": "",
-    "query_processors": [
-        "AdaptiveQueryProcessor"
-    ],
-    "query_mappings": "NOT=true,NOT_CHAR=-",
-    "result_processor": "",
-    "result_processors": [
-        "MappingResultProcessor"
-    ],
-    "response_mappings": "",
-    "result_mappings": "title=summary,date_published=resource.createdDateTime,author=resource.from.emailAddress.address,NO_PAYLOAD",
-    "results_per_query": 10,
-    "credentials": "",
-    "eval_credentials": "",
-    "tags": [
-        "mgraph"
-    ]
-}
