--- pytest.ini
@@ -1,8 +1,0 @@
-[pytest]
-DJANGO_SETTINGS_MODULE = swirl_server.settings
-python_files = tests.py test_*.py *_tests.py
-addopts = --verbose
-log_cli = 1
-log_format = %(asctime)s %(levelname)s %(message)s
-log_date_format = %Y-%m-%d %H:%M:%S
-log_cli_level = ERROR
