--- requirements-test.txt
@@ -1,3 +1,0 @@
-pytest
-pytest-django
-responses
