+++ requirements.txt
@@ -20,7 +20,6 @@ bs4
 google-cloud-bigquery
 opensearch-py
 openai
-msal
