--- swirl/authenticators/__init__.py
@@ -1,3 +1,0 @@
-from swirl.authenticators.authenticator import Authenticator
-from swirl.authenticators.microsoft import Microsoft
-
