+++ swirl/banner.py
@@ -10,13 +10,13 @@ class bcolors:
     ENDC = '\033[0m'
     BOLD = '\033[1m'
 
-SWIRL_VERSION = '1.11'
+SWIRL_VERSION = '1.10'
 
-SWIRL_BANNER_TEXT = "__S_W_I_R_L__1_._1_1____________________________________________________________"
+SWIRL_BANNER_TEXT = "__S_W_I_R_L__1_._1_0____________________________________________________________"
 SWIRL_BANNER = f'{bcolors.BOLD}{SWIRL_BANNER_TEXT}{bcolors.ENDC}'
 
-#############################################
-
+#############################################    
+    
 if __name__ == "__main__":
     print(SWIRL_BANNER)
