+++ swirl/connectors/__init__.py
@@ -4,18 +4,11 @@
 '''
 
 from swirl.connectors.requestsget import RequestsGet
-from swirl.connectors.requestspost import RequestsPost
 from swirl.connectors.sqlite3 import Sqlite3
 from swirl.connectors.elastic import Elastic
 from swirl.connectors.opensearch import OpenSearch
 from swirl.connectors.bigquery import BigQuery
 from swirl.connectors.chatgpt import ChatGPT
-from swirl.connectors.microsoft_graph import M365OutlookMessages
-from swirl.connectors.microsoft_graph import M365OneDrive
-from swirl.connectors.microsoft_graph import M365OutlookCalendar
-from swirl.connectors.microsoft_graph import M365SharePointSites
-from swirl.connectors.microsoft_graph import MicrosoftTeams
-
 
 # uncomment the line below to enable PostgreSQL
 # from swirl.connectors.postgresql import PostgreSQL
