+++ swirl/models.py
@@ -22,17 +22,6 @@ class SearchProvider(models.Model):
     date_updated = models.DateTimeField(auto_now=True)
     active = models.BooleanField(default=True)
     default = models.BooleanField(default=True)
-    AUTHENTICATOR_CHOICES = [
-        ('Microsoft', 'Microsoft Authentication')
-    ]
-    authenticator = models.CharField(max_length=200, default='', choices=AUTHENTICATOR_CHOICES)
-    CONNECTORS_AUTHENTICATORS = dict({
-        'M365OutlookMessages': 'Microsoft',
-        'M365OneDrive': 'Microsoft',
-        'M365OutlookCalendar': 'Microsoft',
-        'M365SharePointSites': 'Microsoft',
-        'MicrosoftTeams': 'Microsoft',
-    })
     CONNECTOR_CHOICES = [
         ('ChatGPT', 'ChatGPT Query String'),
         ('RequestsGet', 'HTTP/GET returning JSON'),
@@ -41,14 +30,8 @@ class SearchProvider(models.Model):
         # Uncomment the line below to enable PostgreSQL
         # ('PostgreSQL', 'PostgreSQL'),
         ('BigQuery', 'Google BigQuery'),
-        ('Sqlite3', 'Sqlite3'),
-        ('M365OutlookMessages', 'M365 Outlook Messages'),
-        ('M365OneDrive', 'M365 One Drive'),
-        ('M365OutlookCalendar', 'M365 Outlook Calendar'),
-        ('M365SharePointSites', 'M365 SharePoint Sites'),
-        ('MicrosoftTeams', 'Microsoft Teams'),
+        ('Sqlite3', 'Sqlite3')
     ]
-
     connector = models.CharField(max_length=200, default='RequestsGet', choices=CONNECTOR_CHOICES)
     url = models.CharField(max_length=2048, default=str, blank=True)
     query_template = models.CharField(max_length=2048, default='{url}?q={query_string}', blank=True)
@@ -72,7 +55,6 @@ class SearchProvider(models.Model):
     result_processors = models.JSONField(default=getSearchProviderResultProcessorsDefault, blank=True)
     result_mappings = models.CharField(max_length=2048, default=str, blank=True)
     results_per_query = models.IntegerField(default=10)
-    eval_credentials = models.CharField(max_length=100, default=str, blank=True)
     credentials = models.CharField(max_length=512, default=str, blank=True)
     tags = models.JSONField(default=list)
 
@@ -193,23 +175,3 @@ class Result(models.Model):
     def __str__(self):
         signature = str(self.id) + ':' + str(self.search_id) + ':' + str(self.searchprovider)
         return signature
-
-class QueryTransform(models.Model) :
-    id = models.BigAutoField(primary_key=True)
-    # DN : TODO name should be unique.
-    name = models.CharField(max_length=255)
-    owner = models.ForeignKey('auth.User', on_delete=models.CASCADE)
-    shared = models.BooleanField(default=False)
-    date_created = models.DateTimeField(auto_now_add=True)
-    date_updated = models.DateTimeField(auto_now=True)
-    QUERY_TRASNSFORM_TYPE_CHOICES = [
-        ('rewrite', 'Rewrite'),
-        ('synonym', 'Synonym' ),
-        ('bag', 'Synonym Bag' )
-    ]
-    qrx_type =  models.CharField(max_length=64, default='', choices=QUERY_TRASNSFORM_TYPE_CHOICES)
-    config_content = models.TextField()
-    class Meta:
-        unique_together = [
-            ('name', 'qrx_type'),
-        ]
