+++ swirl/nltk.py
@@ -23,4 +23,3 @@ from nltk.stem import PorterStemmer
 ps = PorterStemmer()
 
 from nltk.tokenize import sent_tokenize
-from nltk.tokenize import word_tokenize
