+++ swirl/processors/__init__.py
@@ -10,6 +10,5 @@ from swirl.processors.dedupe import *
 from swirl.processors.relevancy import *
 from swirl.processors.spellcheck_query import *
 from swirl.processors.chatgpt_query import *
-from swirl.processors.transform_query_processor import *
 
 # Add new processors here!
