+++ swirl/tasks.py
@@ -33,11 +33,11 @@ for t in SWIRL_OBJECT_LIST:
 ##################################################
 
 @shared_task(name='federate', ignore_result=True)
-def federate_task(search_id, provider_id, provider_connector, update, session):
+def federate_task(search_id, provider_id, provider_connector, update):
     logger.info(f"{module_name}: federate_task: {search_id}_{provider_id}_{provider_connector} update: {update}")
     try:
         connector = eval(provider_connector, {"provider_connector": provider_connector, "__builtins__": None}, SWIRL_OBJECT_DICT)(provider_id, search_id, update)
-        connector.federate(session)
+        connector.federate()
     except NameError as err:
         message = f'Error: NameError: {err}'
         logger.error(f'{module_name}: {message}')
@@ -48,40 +48,36 @@ def federate_task(search_id, provider_id, provider_connector, update, session):
 
 ##################################################
 
+from swirl.search import search
 
 @shared_task(name='search', ignore_result=True)
-def search_task(search_id, session):
-    from swirl.search import search
-
+def search_task(search_id):
     logger.info(f"{module_name}: search_task: {search_id}")
-    return search(search_id, session)
+    return search(search_id)
 
 ##################################################
 
+from swirl.search import rescore
 
 @shared_task(name='rescore', ignore_result=True)
 def rescore_task(search_id):
-    from swirl.search import rescore
-
     logger.info(f"{module_name}: rescore_task: {search_id}")
     return rescore(search_id)
 
 ##################################################
 
+from swirl.expirer import expirer
 
 @shared_task(name='expirer')
 def expirer_task():
-    from swirl.expirer import expirer
-
     logger.info(f"{module_name}: expirer()")
     return expirer()
 
 ##################################################
 
+from swirl.subscriber import subscriber
 
 @shared_task(name='subscriber')
 def subscriber_task():
-    from swirl.subscriber import subscriber
-
     logger.info(f"{module_name}: subscriber()")
     return subscriber()
