--- swirl/templates/authenticators.html
@@ -1,27 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-
-    <h2>Authenticators:</h2>
-    <table>
-        <tbody>
-            {% for a in authenticators %}
-                <form method="post">
-                    {% csrf_token %}
-                    <input type="hidden" name="authenticator_name" value="{{ a.name }}"></input>
-                    <tr style="display: flex; flex-direction: row; width: 100%;">
-                        <td>{{ a.name }}</td>
-                        <td style="margin-left: 50px">
-                            <button type="submit" name="{{ a.name }}">
-                                Refresh token
-                            </button>
-                        </td>
-                    </tr>
-                </form>
-            {% endfor %}
-        </tbody>
-    </table>
-    <br/><br/>
-
-{% endblock %}
-
