+++ swirl/templates/index.html
@@ -3,14 +3,10 @@
 {% block title %}Home{% endblock %}
 
 {% block content %}
-&#10230; <a href="/swirl/authenticators.html" target="_new">Authenticators</a><br/><br/>
-&#10230; <a href="/swirl/search.html" target="_new">Search Form</a><br/><br/>
-&#10230; <a href="/swirl/query_transform_form/" target="_new">Add Query Transforms</a><br/>
 {% load env_extras %}
 &#10230; <a href="{% get_search_form_url %}" target="_new">Search Form</a><br/><br/>
 APIs:<br/>
 &#10230; <a href="/swirl/searchproviders/" target="_new">SearchProviders</a><br/>
-&#10230; <a href="/swirl/querytransforms/" target="_new">QueryTransforms</a><br/>
 &#10230; <a href="/swirl/search/" target="_new">Search</a><br/>
 &#10230; <a href="/swirl/results/" target="_new">Results</a><br/>
 &#10230; <a href="/swirl/swagger-ui/" target="_new">Swagger</a><br/><br/>
