--- swirl/templates/query_transform.html
@@ -1,12 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-  <h2>Query Transform</h2>
-  <form method="post" enctype="multipart/form-data">
-    {% csrf_token %}
-    <p>{{ form.name.label_tag }}: {{ form.name }}</p>
-    <p>{{ form.content_type.label_tag }}: {{ form.content_type }}</p>
-    <p>{{ form.file.label_tag }}: {{ form.file }}</p>
-    <button type="submit">Upload</button>
-  </form>
-{% endblock %}
