+++ swirl_server/settings.py
@@ -177,7 +177,6 @@ CELERY_TASK_SERIALIZER = 'json'
 CELERY_RESULT_SERIALIZER = 'json'
 CELERY_ACCEPT_CONTENT = ['json']
 CELERY_RESULT_BACKEND='rpc://'
-CELERY_TASK_ALWAYS_EAGER = False
 
 CELERY_WORKER_CANCEL_LONG_RUNNING_TASKS_ON_CONNECTION_LOSS = False
 
@@ -247,8 +246,3 @@ if 'SWIRL_SEARCH_FORM_URL' in env:
 
 if 'OPENAI_API_KEY' in env:
     OPENAI_API_KEY = env('OPENAI_API_KEY')
-
-
-MICROSOFT_CLIENT_ID = ''
-MICROSOFT_CLIENT_SECRET = ''
-MICROSOFT_REDIRECT_URI = 'http://localhost:8000/swirl/microsoft-callback'
